reorganized entire codebase
