+++ client/src/app/api/auth/[...nextauth]/route.ts
