+++ client/src/app/page.css
@@ -8,3 +8,8 @@
 body {
   font-family: 'Inter', sans-serif;
 }
+
+.loading-image{
+  width: 50vw;
+  height: auto;
+}
