+++ client/src/app/page.tsx
@@ -1,15 +1,18 @@
 /* eslint-disable @next/next/no-img-element */
 "use client";
 
-import Dashboard from "@/pages/Dashboard";
-import Login from "@/pages/Login";
-import CustomHeader from "@/components/Header/CustomHeader";
-import CustomSider from "@/components/Sider/CustomSider";
+import Dashboard from "../pages/Dashboard";
+import Login from "../pages/Login";
+import CustomHeader from "../components/Header/CustomHeader";
+import CustomSider from "../components/Sider/CustomSider";
 import { useSession } from "next-auth/react";
 import { useEffect, useState } from "react";
 import { Layout } from "antd";
 import { Content } from "antd/es/layout/layout";
 import "./page.css";
+import icon from '../../../public/full_logo.png'
+import Image from 'next/image';
+
 
 export default function Home() {
   const [isLoggedIn, setIsLoggedIn] = useState(false);
@@ -36,7 +39,7 @@ export default function Home() {
         height: "100vh",
       }}
     >
-      <img src="./full_logo.png" alt="icon" />
+      <Image src={icon} alt="icon" width={500} height={100} className="loading-image"/>
     </div>
   ) : isLoggedIn ? (
     <Login />
