+++ client/src/common/interfaces.ts
