+++ client/src/components/Header/CustomHeader.tsx
@@ -1,10 +1,10 @@
 import React from "react";
 import { Layout } from "antd";
 import Image from "next/image";
-import icon from "../../../public/full_logo.png";
-import { PageProps } from "@/common/interfaces";
+import icon from '../../../../public/full_logo.png'
+import { PageProps } from "../../common/interfaces";
 import "./CustomHeader.css";
-import { accentBackground, buRed } from "@/common/styles";
+import { accentBackground, buRed } from "../../common/styles";
 import Pfp from "../Pfp";
 
 const { Header } = Layout;
