--- package.json
@@ -1,5 +1,0 @@
-{
-  "dependencies": {
-    "styled-components": "^6.1.11"
-  }
-}
